requirements.txt
attrs==23.1.0
websockets==11.0.2
websocket-client==1.5.1
